requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
exceptiongroup==1.3.0
fastapi==0.115.12
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
locust==2.37.10
locust-cloud==1.23.1
MarkupSafe==3.0.2
msgpack==1.1.1
networkx==2.8.8
numpy==1.23.5
packaging==25.0
pandas==1.5.3
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
scikit-learn==1.1.3
scipy==1.9.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.22.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
gradio